Added session authentication - login and signup + logout

+++ app.py
@@ -1,17 +1,76 @@
-from flask import Flask, render_template, request, redirect
+from flask import Flask, render_template, request, redirect, session
 from datetime import datetime
 import sqlite3
 import database
 
 app = Flask(__name__)
+app.secret_key = "your-secret-key"
 
 def get_connection():
     conn = sqlite3.connect("spendly.db")
     conn.row_factory = sqlite3.Row
     return conn
 
+@app.route('/login', methods = ["GET", "POST"])
+def login():
+    if request.method == "GET":
+        return render_template("usersession.html")
+
+    email = request.form["email"]
+    password = request.form["password"]
+
+    conn = get_connection()
+    
+    user = conn.execute("SELECT * FROM users WHERE email=? AND password=?", (email,password)).fetchone()
+
+    conn.close()
+
+    if not user:
+        return "Invalid credentials"
+    
+    session["user_id"] = user["id"]
+    session["user_name"] = user["name"]
+
+    return redirect("/")
+
+@app.route("/signup", methods=["POST"])
+def signup():
+    name = request.form["name"]
+    email = request.form["email"]
+    password = request.form["password"]
+    confirm = request.form["confirm_password"]
+
+    conn = get_connection()
+
+    if password != confirm:
+        return "Passwords do not match"
+      
+    try:
+        conn.execute(
+            "INSERT INTO users (name, email, password) VALUES (?, ?, ?)",
+            (name, email, password)
+        )
+        conn.commit()
+
+    except:
+        return "User already exists"
+
+    finally:
+        conn.close()
+
+    return redirect('/login')
+
+@app.route("/logout")
+def logout():
+    session.pop("user_id", None)
+    # session.clear()     both do the same work
+    return redirect("/login")
+
 @app.route('/')
 def home():
+    if "user_id" not in session:
+        return redirect('/login')
+    
     current_monthYear = datetime.now().strftime("%B %Y")
     conn = get_connection()
     budget_percentage = 0
@@ -20,33 +79,37 @@ def home():
     SELECT COALESCE(SUM(amount),0)
     FROM transactions
     WHERE transaction_type='income'
-    """).fetchone()[0]
+    AND user_id=?
+    """,(session["user_id"],)).fetchone()[0]
 
     expenses = conn.execute("""
     SELECT COALESCE(SUM(amount),0)
     FROM transactions
     WHERE transaction_type='expense'
-    """).fetchone()[0]
+    AND user_id=?
+    """, (session["user_id"],)).fetchone()[0]
 
     budget_row = conn.execute("""
     SELECT amount
     FROM budget
-    WHERE id=1
-    """).fetchone()
+    WHERE user_id=?
+    """, (session["user_id"],)).fetchone()
 
     transactions = conn.execute("""
     SELECT *
     FROM transactions
+    WHERE user_id=?
     ORDER BY created_at DESC
     LIMIT 5
-    """).fetchall()
+    """, (session["user_id"],)).fetchall()
 
     category_data = conn.execute("""
     SELECT category, SUM(amount) as total
     FROM transactions
     WHERE transaction_type='expense'
+    AND user_id=?
     GROUP BY category
-    """).fetchall()
+    """, (session["user_id"],)).fetchall()
 
     category_data = [dict(row) for row in category_data]
 
@@ -55,13 +118,14 @@ def home():
         SUM(amount) as total
     FROM transactions
     WHERE transaction_type='expense'
+    AND user_id=?
     GROUP BY category
     ORDER BY total DESC
-    """).fetchall()
+    """, (session["user_id"],)).fetchall()
 
     monthly_data = [dict(row) for row in monthly_data]
 
-    budget = budget_row["amount"] if budget_row else 0
+    budget = float(budget_row["amount"]) if budget_row else 0.0
     budget = float(budget)
 
     conn.close()
@@ -112,6 +176,9 @@ def home():
 
 @app.route('/add-income', methods=['POST'])
 def add_income():
+    if "user_id" not in session:
+        return redirect("/login")
+    
     amount = request.form.get('amount')
     category = request.form.get('category')
     amount = float(amount)
@@ -121,20 +188,22 @@ def add_income():
     conn.execute(
         """
         INSERT INTO transactions
-        (transaction_type, amount, category)
-        VALUES (?, ?, ?)
+        (user_id, transaction_type, amount, category)
+        VALUES (?, ?, ?, ?)
         """,
-        ("income", amount, category)
+        (session["user_id"], "income", amount, category)
     )
 
     conn.commit()
     conn.close()
 
-    # later: store in DB
     return redirect('/')
 
 @app.route('/add-expense', methods=['POST'])
 def add_expense():
+    if "user_id" not in session:
+        return redirect("/login")
+
     amount = request.form.get('amount')
     category = request.form.get('category')
     amount = float(amount)
@@ -144,10 +213,10 @@ def add_expense():
     conn.execute(
         """
         INSERT INTO transactions
-        (transaction_type, amount, category)
-        VALUES (?, ?, ?)
+        (user_id, transaction_type, amount, category)
+        VALUES (?, ?, ?, ?)
         """,
-        ("expense", amount, category)
+        (session["user_id"], "expense", amount, category)
     )
 
     conn.commit()
@@ -158,6 +227,9 @@ def add_expense():
 
 @app.route('/setBudget', methods = ['POST'])
 def setBudget():
+    if "user_id" not in session:
+        return redirect("/login")
+
     amount = request.form.get('amount')
     amount = float(amount)
 
@@ -166,10 +238,9 @@ def setBudget():
     conn.execute(
         """
         INSERT OR REPLACE INTO budget
-        (id, amount)
-        VALUES (1, ?)
-        """,
-        (amount,)
+        (user_id, amount)
+        VALUES (?, ?)
+        """, (session["user_id"], amount)
     )
 
     conn.commit()
@@ -179,6 +250,9 @@ def setBudget():
 
 @app.route('/edit/<int:id>')
 def edit_transaction(id):
+    if "user_id" not in session:
+        return redirect("/login")
+
     conn = get_connection()
 
     transaction = conn.execute(
@@ -186,8 +260,9 @@ def edit_transaction(id):
         SELECT *
         FROM transactions
         WHERE id=?
+        AND user_id=?
         """,
-        (id,)
+        (id, session["user_id"])
     ).fetchone()
 
     conn.close()
@@ -199,11 +274,14 @@ def edit_transaction(id):
 
 @app.route('/delete/<int:id>', methods = ['POST'])
 def delete_transaction(id):
+    if "user_id" not in session:
+        return redirect("/login")
+
     conn = get_connection()
 
     conn.execute(
-        "DELETE FROM transactions WHERE id=?",
-        (id,)
+        "DELETE FROM transactions WHERE id=? AND user_id=?",
+        (id, session["user_id"])
     )
 
     conn.commit()
@@ -213,6 +291,9 @@ def delete_transaction(id):
 
 @app.route('/update/<int:id>', methods=['POST'])
 def update_transaction(id):
+    if "user_id" not in session:
+        return redirect("/login")
+
     amount = request.form.get("amount")
     category = request.form.get("category")
 
@@ -222,15 +303,17 @@ def update_transaction(id):
         """
         UPDATE transactions
         SET amount=?, category=?
-        WHERE id=?
+        WHERE id=? AND user_id=?
         """,
-        (amount, category, id)
+        (amount, category, id, session["user_id"])
     )
 
     conn.commit()
     conn.close()
 
     return redirect('/')
+
+# print(session.get("user_id"))
 
 if __name__ == "__main__":
     app.run(debug=True)
